requirements.txt
grpcio==1.64.0
numpy==1.26.4
Oasys.gRPC==21.0.0
Oasys.PRIMER==21.0.0
pandas==2.2.2
protobuf==5.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
wheel==0.43.0
